requirements.txt
gspread==5.12.0
oauth2client==4.1.3
google-api-python-client==2.111.0
google-auth==2.27.0
cohere==5.9.2
matplotlib==3.8.2
pandas==2.1.4
gspread-formatting==1.2.0
python-dotenv==1.0.1